requirements.txt
absl-py==2.1.0
altair==5.3.0
attrs==23.2.0
blinker==1.8.2
Brotli 
cachetools==5.3.3
opencv-python-headless
certifi 
charset-normalizer 
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock 
fonttools==4.53.0
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.64.0
idna 
intel-openmp==2021.4.0
Jinja2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe
matplotlib==3.9.0
mdurl==0.1.2
mkl==2021.4.0
mpmath
mss==9.0.1
networkx
numpy
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
PySocks 
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML 
referencing==0.35.1
requests
rich==13.7.1
rpds-py==0.18.1
scipy==1.13.1
seaborn==0.13.2
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
streamlit==1.35.0
sympy
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
typing_extensions
tzdata==2024.1
ultralytics==8.2.27
urllib3 
watchdog==4.0.1
Werkzeug==3.0.3
wheel==0.43.0
win-inet-pton
